Fixed symlink notifications on Linux
The linux code used to ignore events on the files that os.Stat reports as
non existing. This ignores the CREATEion of broken links, but monitoring
those can be useful, for example /proc/<pid>/fd/ contains fake links
for all sockets.

The patch simply replaces os.Stat with os.Lstat to solve the issue.

I updated the fsnotify_symlink_test.go to what I consider the correct
behavior but unfortunately couldn't test under windows, so I'm not sure
if it breaks the test there or not.