kqueue: fix Remove-logic
`os.Lstat` does not return `error` when a file/directory exists,
so the code path  to check for newly created files after a REMOVE
on BSD was never executed.

Also, if the watched entry is a file, it has been removed from the watcher, so do a new-file-check for them as well.

This commit fixes issues on TextMate on OS X and others that uses `exchangedata` to do atomic saves.

See #17. closes #111.