Document and test removing watched directory on Windows
On Windows deleting the watched directive gives undefined results and
delivering events for the files in the directory is not guaranteed. I
can reproduce this even with a very simple example in Python, and it's
not an fsnotify bug as far as I can see.

There's nothing we can do about this, so update the documentation and
tweak the test so we don't get intermittent test failures.

On other platforms it does report all files.