Expose and document Supports, Unportable*
Only implemented on Linux/inotify for now. I had hoped to finish some
more platforms before exporting this, but I accidentally deleted my work
on that a few months ago and haven't had the energy to restart it.

Lots of people want it, and there's are a few "forks" of fsnotify which
export this. So just add it in the release.