Make BufferedWatcher buffered again
The size parameter didn't actually get passed to the make() call;
regression from #632. Also make sure the default buffer size is 50 on
Windows again.

There is no need for a separate newBackend() and newBufferedBackend() in
the first place, as the buffer size just controls the channel buffer
size which is now done in the generic fsnotify.go, so refactor to remove
that.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>