Reduce intermittent "bad file descriptor" errors on kqueue
It's a long-standing problem that some tests intermittently fail on the
CI with this, but with the new TestRace/remove_self it would be fairly
consistently reproducible.

Seems what would happen is that it did "fd := w.path[path]", and that
this could return 0 as that path had already been removed. That would
close fd 0 (stdin), fd 0 then got re-used for the next test, which would
work for a bit until it didn't and you got an error.

Also ignore some ErrNotExist errors.

This doesn't completely fix it; it still fails on occasion. I want to
rewrite the kqueue backend anyway, and it's been a problem for as long
as I've been involved. So not a priority to completely fix it right now.