Don't add double watches for symlinks
inotify_add_watch() follows symlinks, and returns the current watch
descriptor when adding a patch twice. So when doing "watch dir" and
"watch link" (or reverse) second watch is basically a no-op; yet it's
still registered as a "separate" watch, and would panic on removing the
second.

The solution is to make the second Add() a no-op. This is also what
kqueue does, and what happens if you watch the same path twice.

On illumos watching a symlink currently means registering double
watches; this is a separate bug that should be fixed.

Fixes #652
Fixes #662