Fix relative links on kqueue
Also only resolve symlinks explicitly given in Add()/AddWith(), and not
those added "internally" as part of a listdir. Also don't ignore errors
from os.Readlink() and os.Lstat; this was added in 2012 with bf36090 to
ignore unresolvable symlinks when listing a directory, but we're no
longer calling lstat() on the result of a link unless it's explicitly
passed with Add().

Fixes #661