Simplify inotify backend locking
Change locking to be less fine-grained: just lock the watcher while
we're processing an event, so we're operating on a consistent
"snapshot".

This is basically how it worked before 16df002. Moving some of
bookkeeping to a separate "helper struct" was a good idea; doing the
locking in that struct wasn't.