Fix missing kqueue notifications for create + write
Fixes #717

On macOS, a watcher can miss notifications for written data when
watching a directory, and a file is removed + re-created with content.
Though the watcher gets a CREATE event, the file may not have the
content yet, and the WRITE notification is never received.

With kqueue, each new file requires a new watch, but it's possible that
a write to the file happens before the watch is set up. This results in
missed WRITEs. However, even if the caller checks the content on CREATE,
they can miss written data, as the watch is setup *after* sending the
CREATE event.

Instead, if we flip the watch to be setup first, then a caller on CREATE
will either see the data, or will receive a WRITE event when the data
is written.