Use Go 1.25 in CI, move FreeBSD test runner to GitHub actions
The Cirrus one isn't that faster, and breaks more often as the images
change, and I'm confused by the UI every time. It used to be that the
GitHub vmactions ones were very very slow, but that's no longer case
since they switched from macOS to Linux hosts. So, just use GitHub
Actions.

Remove some exceptions we previously had for Dragonfly BSD; seems that
got updated/improved?

Skip some of the recursive tests for Windows. These were always a tad
flaky, but seem to reliably fail with Go 1.25. None of this is exposed
yet – need to look in to it.